+++ data/src/window.rs
@@ -1,0 +1,69 @@
+use std::io;
+use std::path::PathBuf;
+
+use serde::{Deserialize, Serialize};
+
+use crate::environment;
+
+pub use self::size::Size;
+pub use self::position::Position;
+
+pub mod size;
+pub mod position;
+
+
+#[derive(Debug, Default, Clone, Copy, Serialize, Deserialize)]
+pub struct Window {
+    pub position: Option<Position>,
+    pub size: Size,
+}
+
+impl Window {
+    pub fn update(self, event: Event) -> Self {
+        match event {
+            Event::Moved(position) => Self { position: Some(position), ..self },
+            Event::Resized(size) => Self { size, ..self },
+        }
+    }
+
+    pub fn load() -> Result<Self, Error> {
+        let path = path()?;
+
+        let bytes = std::fs::read(path)?;
+
+        Ok(serde_json::from_slice(&bytes)?)
+    }
+
+    pub async fn save(self) -> Result<(), Error> {
+        let path = path()?;
+
+        let bytes = serde_json::to_vec(&self)?;
+        tokio::fs::write(path, &bytes).await?;
+
+        Ok(())
+    }
+}
+
+fn path() -> Result<PathBuf, Error> {
+    let parent = environment::data_dir();
+
+    if !parent.exists() {
+        std::fs::create_dir_all(&parent)?;
+    }
+
+    Ok(parent.join("window.json"))
+}
+
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error(transparent)]
+    Serde(#[from] serde_json::Error),
+    #[error(transparent)]
+    Io(#[from] io::Error),
+}
+
+#[derive(Debug, Clone, Copy)]
+pub enum Event {
+    Moved(Position),
+    Resized(Size),
+}